cleaned up sshfs support and logging

+++ bridgy/sshfs.py
@@ -7,15 +7,24 @@ import logging
 from config import Config
 import ssh
 
+logger = logging.getLogger(__name__)
 
-# [[source, mountpoint, filesystem, options, p1, p2],...]
+def _getSystemMounts():
+    lines = [line.strip("\n").split(" ") for line in open("/etc/mtab", "r").readlines()]
+    sshfsMounts = [mp for src, mp, fs, opt, p1, p2 in lines if fs == "fuse.sshfs"]
+    return sshfsMounts
+
+
 def getMounts():
-    try:
-        lines = [line.strip("\n").split(" ")
-                 for line in open("/etc/mtab", "r").readlines()]
-        return [mount for mount in lines if mount[2] == "fuse.sshfs"]
-    except:
-        print "Could not read mtab"
+    systemMounts = set(_getSystemMounts())
+    possibleOwnedMounts = set([os.path.join(Config.mountRootDir, d) for d in os.listdir(Config.mountRootDir)])
+
+    ownedMounts = systemMounts & possibleOwnedMounts
+    return list(ownedMounts)
+
+def isMounted(instance):
+    mp = getMountpoint(instance)
+    return mp in getMounts()
 
 def getMountpoint(instance):
     return os.path.join(Config.mountRootDir, '%s@%s'%(instance.name, instance.address))
@@ -25,21 +34,17 @@ def mount(instance, remotedir):
     if not os.path.exists(mountpoint):
         os.mkdir(mountpoint)
 
-    currentMounts = [mp for src, mp, fs, opts, p1, p2 in getMounts()]
-    if mountpoint in currentMounts:
-        print "Already mounted at %s" % mountpoint
+    if mountpoint in getMounts():
+        logger.warn("Already mounted at %s" % mountpoint)
         sys.exit(1)
 
     rc = os.system(ssh.SshfsCommand(instance, remotedir, mountpoint))
     if rc == 0:
-        print "%s mounted as %s" % (instance, mountpoint)
-    else:
-        logging.getLogger().error("Failed to mount instance {} at {}".format(instance, mountpoint))
-        os.rmdir(mountpoint)
+        return True
 
-def umountAll():
-    for src, mp, fs, opts, p1, p2 in getMounts():
-        umount(mountpoint=mp)
+    logger.error("Failed to mount instance {} at {}".format(instance, mountpoint))
+    os.rmdir(mountpoint)
+    return False
 
 def umount(mountpoint=None, instance=None):
     if mountpoint == None and instance == None or mountpoint != None and instance != None:
@@ -48,18 +53,7 @@ def umount(mountpoint=None, instance=None):
     if instance:
         mountpoint = getMountpoint(instance)
 
-    if os.path.exists(mountpoint):
-        os.system("fusermount -u %s" % mountpoint)
+    if os.path.exists(mountpoint) and os.system("fusermount -u %s" % mountpoint) == 0:
         os.rmdir(mountpoint)
-        print 'unmounted'
-    else:
-        print 'unable to unmount'
-
-def list():
-    mounts = getMounts()
-
-    if len(mounts) > 0:
-        for src, mp, fs, opts, p1, p2 in mounts:
-            print mp
-    else:
-        print "No mounts."
+        return True
+    return False
